Omit empty children by default from hash/json
Empty children is annoying and clutters up the json output.
Additionally it is misleading because in actuality the item _may_ have
children, but due to the query none of them were returned.

If you want to see the children regardless of emptiness, you can pass
a flag.

Fixes CNVS-26736

Test Plan:

    - Grab the list of Authorities:
        - a = handle.standards.authorities
    - Convert it to json and pretty print it:
        - JSON.pretty_generate(a.map(&:to_h))
    - Observe that the authorities do not have any children objects

Change-Id: I12dbd8d606e7b095d83727b6e2e1efc5e6929e7b