Prune orphans from Standards Trees
When searching for all standards in an authority, we sometimes get back
some standards that have a parent guid, but the standard owning that
guid is not in the list of standards.  This means that these standards
are orhpans.  They aren't roots because they have a parents, but the
parents aren't part of this authority.

We have to do _something_ with these orphans.  They don't fit into the
tree hierarchy anywhere.  For now, we will prune them from the tree but
save them for later retrieval.  This way the client can decide how to
deal with these depending on the needs of their application.

Fixes CNVS-26997

Test Plan:

    - Build an authority tree with the Utah standards.  NOTE: This
      requires you to use a subscription that contains the Utah
      standards
      - utah_tree = ab_handle.standards.authority_tree("Utah")
    - Observe that there are no errors, and that there is at least one
      orphan present
      - utah_tree.orphans.count > 0

Change-Id: I8ebdba20f1a750df34b39341c4616489bdce4917